Tuple annotation

+++ 0x00-python_variable_annotations/7-to_kv.py
@@ -9,4 +9,4 @@ from typing import Union, Tuple
 
 
 def to_kv(k: str, v: Union[int, float]) -> Tuple[str, float]:
-    return (k, float(v * v))
+    return (k, (v * v))
